+++ boxes/boxes/vite/src/config.ts
@@ -15,7 +15,7 @@ import { KeyStore } from "@aztec/key-store";
 import { PrivateKernelProver } from "@aztec/circuit-types";
 import { L2TipsStore } from "@aztec/kv-store/stores";
 import { createStore } from "@aztec/kv-store/indexeddb";
-import { BbWasmAsyncPrivateKernelProver } from "@aztec/bb-prover/wasm_async";
+import { BBWASMLazyPrivateKernelProver } from "@aztec/bb-prover/wasm/lazy";
 
 process.env = Object.keys(import.meta.env).reduce((acc, key) => {
   acc[key.replace("VITE_", "")] = import.meta.env[key];
@@ -39,7 +39,7 @@ export class PrivateEnv {
     config.dataDirectory = "pxe";
     config.proverEnabled = true;
     const aztecNode = await createAztecNodeClient(this.nodeURL);
-    const proofCreator = new BbWasmAsyncPrivateKernelProver(16);
+    const proofCreator = new BBWASMLazyPrivateKernelProver(16);
     this.pxe = await this.createPXEService(aztecNode, config, proofCreator);
     const encryptionPrivateKey = deriveMasterIncomingViewingSecretKey(
       this.secretKey,
