+++ yarn-project/bb-prover/src/prover/bb_private_kernel_prover.ts
@@ -1,7 +1,7 @@
 import { type PrivateKernelProver, type PrivateKernelSimulateOutput } from '@aztec/circuit-types';
-import { type CircuitSimulationStats, type CircuitWitnessGenerationStats } from '@aztec/circuit-types/stats';
+import { CircuitSimulationStats } from '@aztec/circuit-types/stats';
 import {
-  type ClientIvcProof,
+  ClientIvcProof,
   type PrivateKernelCircuitPublicInputs,
   type PrivateKernelInitCircuitPrivateInputs,
   type PrivateKernelInnerCircuitPrivateInputs,
@@ -9,107 +9,39 @@ import {
   type PrivateKernelTailCircuitPrivateInputs,
   type PrivateKernelTailCircuitPublicInputs,
 } from '@aztec/circuits.js';
-import { runInDirectory } from '@aztec/foundation/fs';
-import { type Logger, createLogger } from '@aztec/foundation/log';
+import { createLogger } from '@aztec/foundation/log';
 import { Timer, elapsed } from '@aztec/foundation/timer';
 import {
-  ClientCircuitArtifacts,
-  type ClientProtocolArtifact,
-  convertPrivateKernelInitInputsToWitnessMap,
-  convertPrivateKernelInitOutputsFromWitnessMap,
-  convertPrivateKernelInnerInputsToWitnessMap,
-  convertPrivateKernelInnerOutputsFromWitnessMap,
-  convertPrivateKernelResetInputsToWitnessMap,
-  convertPrivateKernelResetOutputsFromWitnessMap,
-  convertPrivateKernelTailForPublicOutputsFromWitnessMap,
-  convertPrivateKernelTailInputsToWitnessMap,
-  convertPrivateKernelTailOutputsFromWitnessMap,
-  convertPrivateKernelTailToPublicInputsToWitnessMap,
-  executeInit,
-  executeInner,
-  executeReset,
-  executeTail,
-  executeTailForPublic,
+  convertPrivateKernelInitInputsToWitnessMapWithAbi,
+  convertPrivateKernelInitOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelInnerInputsToWitnessMapWithAbi,
+  convertPrivateKernelInnerOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelResetInputsToWitnessMapWithAbi,
+  convertPrivateKernelResetOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelTailForPublicOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelTailInputsToWitnessMapWithAbi,
+  convertPrivateKernelTailOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelTailToPublicInputsToWitnessMapWithAbi,
+  executeInitWithArtifact,
+  executeInnerWithArtifact,
+  executeResetWithArtifact,
+  executeTailForPublicWithArtifact,
+  executeTailWithArtifact,
   getPrivateKernelResetArtifactName,
   maxPrivateKernelResetDimensions,
-} from '@aztec/noir-protocol-circuits-types/client/bundle';
+} from '@aztec/noir-protocol-circuits-types/client';
+import { ArtifactProvider, type ClientProtocolArtifact } from '@aztec/noir-protocol-circuits-types/types';
 import { ClientCircuitVks } from '@aztec/noir-protocol-circuits-types/vks';
-import { WASMSimulatorWithBlobs } from '@aztec/simulator';
-import { type NoirCompiledCircuit } from '@aztec/types/noir';
+import { WASMSimulator } from '@aztec/simulator/client';
+import { NoirCompiledCircuit } from '@aztec/types/noir';
 
-import { encode } from '@msgpack/msgpack';
-import { serializeWitness } from '@noir-lang/noirc_abi';
-import { type WitnessMap } from '@noir-lang/types';
-import { promises as fs } from 'fs';
-import path from 'path';
+import { type WitnessMap } from '@noir-lang/noir_js';
+import { Abi } from '@noir-lang/types';
 
-import { BB_RESULT, computeGateCountForCircuit, executeBbClientIvcProof } from '../bb/execute.js';
-import { type BBConfig } from '../config.js';
-import { mapProtocolArtifactNameToCircuitName } from '../stats.js';
-import { readFromOutputDirectory } from './client_ivc_proof_utils.js';
+export abstract class BBPrivateKernelProver implements PrivateKernelProver {
+  protected simulator = new WASMSimulator();
 
-/**
- * This proof creator implementation uses the native bb binary.
- * This is a temporary implementation until we make the WASM version work.
- * TODO(#7368): this class grew 'organically' aka it could use a look at its resposibilities
- */
-export class BBNativePrivateKernelProver implements PrivateKernelProver {
-  private simulator = new WASMSimulatorWithBlobs();
-
-  private constructor(
-    private bbBinaryPath: string,
-    private bbWorkingDirectory: string,
-    private skipCleanup: boolean,
-    private log = createLogger('bb-prover:native'),
-  ) {}
-
-  public static async new(config: BBConfig, log?: Logger) {
-    await fs.mkdir(config.bbWorkingDirectory, { recursive: true });
-    return new BBNativePrivateKernelProver(config.bbBinaryPath, config.bbWorkingDirectory, !!config.bbSkipCleanup, log);
-  }
-
-  private async _createClientIvcProof(
-    directory: string,
-    acirs: Buffer[],
-    witnessStack: WitnessMap[],
-  ): Promise<ClientIvcProof> {
-    // TODO(#7371): Longer term we won't use this hacked together msgpack format
-    // and instead properly create the bincode serialization from rust
-    await fs.writeFile(path.join(directory, 'acir.msgpack'), encode(acirs));
-    await fs.writeFile(
-      path.join(directory, 'witnesses.msgpack'),
-      encode(witnessStack.map(map => serializeWitness(map))),
-    );
-    const provingResult = await executeBbClientIvcProof(
-      this.bbBinaryPath,
-      directory,
-      path.join(directory, 'acir.msgpack'),
-      path.join(directory, 'witnesses.msgpack'),
-      this.log.info,
-    );
-
-    if (provingResult.status === BB_RESULT.FAILURE) {
-      this.log.error(`Failed to generate client ivc proof`);
-      throw new Error(provingResult.reason);
-    }
-
-    const proof = await readFromOutputDirectory(directory);
-
-    this.log.info(`Generated IVC proof`, {
-      duration: provingResult.durationMs,
-      eventName: 'circuit-proving',
-    });
-
-    return proof;
-  }
-
-  async createClientIvcProof(acirs: Buffer[], witnessStack: WitnessMap[]): Promise<ClientIvcProof> {
-    this.log.info(`Generating Client IVC proof`);
-    const operation = async (directory: string) => {
-      return await this._createClientIvcProof(directory, acirs, witnessStack);
-    };
-    return await this.runInDirectory(operation);
-  }
+  constructor(protected artifactProvider: ArtifactProvider, protected log = createLogger('bb-prover')) {}
 
   public async generateInitOutput(
     inputs: PrivateKernelInitCircuitPrivateInputs,
@@ -117,15 +49,15 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
     return await this.generateCircuitOutput(
       inputs,
       'PrivateKernelInitArtifact',
-      convertPrivateKernelInitInputsToWitnessMap,
-      convertPrivateKernelInitOutputsFromWitnessMap,
+      convertPrivateKernelInitInputsToWitnessMapWithAbi,
+      convertPrivateKernelInitOutputsFromWitnessMapWithAbi,
     );
   }
-
   public async simulateInit(
     privateInputs: PrivateKernelInitCircuitPrivateInputs,
   ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    const [duration, result] = await elapsed(() => executeInit(privateInputs));
+    const artifact = await this.artifactProvider.getSimulatedClientCircuitArtifactByName('PrivateKernelInitArtifact');
+    const [duration, result] = await elapsed(() => executeInitWithArtifact(privateInputs, artifact));
     this.log.debug(`Simulated private kernel init`, {
       eventName: 'circuit-simulation',
       circuitName: 'private-kernel-init',
@@ -142,15 +74,16 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
     return await this.generateCircuitOutput(
       inputs,
       'PrivateKernelInnerArtifact',
-      convertPrivateKernelInnerInputsToWitnessMap,
-      convertPrivateKernelInnerOutputsFromWitnessMap,
+      convertPrivateKernelInnerInputsToWitnessMapWithAbi,
+      convertPrivateKernelInnerOutputsFromWitnessMapWithAbi,
     );
   }
 
   public async simulateInner(
     privateInputs: PrivateKernelInnerCircuitPrivateInputs,
   ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    const [duration, result] = await elapsed(() => executeInner(privateInputs));
+    const artifact = await this.artifactProvider.getSimulatedClientCircuitArtifactByName('PrivateKernelInnerArtifact');
+    const [duration, result] = await elapsed(() => executeInnerWithArtifact(privateInputs, artifact));
     this.log.debug(`Simulated private kernel inner`, {
       eventName: 'circuit-simulation',
       circuitName: 'private-kernel-inner',
@@ -169,17 +102,20 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
     return await this.generateCircuitOutput(
       variantInputs,
       artifactName,
-      variantInputs => convertPrivateKernelResetInputsToWitnessMap(variantInputs, artifactName),
-      output => convertPrivateKernelResetOutputsFromWitnessMap(output, artifactName),
+      convertPrivateKernelResetInputsToWitnessMapWithAbi,
+      convertPrivateKernelResetOutputsFromWitnessMapWithAbi,
     );
   }
 
   public async simulateReset(
     privateInputs: PrivateKernelResetCircuitPrivateInputs,
   ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
+    const artifact = await this.artifactProvider.getSimulatedClientCircuitArtifactByName(
+      getPrivateKernelResetArtifactName(privateInputs.dimensions),
+    );
     const variantPrivateInputs = privateInputs.trimToSizes();
     const [duration, result] = await elapsed(() =>
-      executeReset(variantPrivateInputs, privateInputs.dimensions, privateInputs),
+      executeResetWithArtifact(variantPrivateInputs, artifact, privateInputs),
     );
     this.log.debug(`Simulated private kernel reset`, {
       eventName: 'circuit-simulation',
@@ -201,15 +137,15 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
       return await this.generateCircuitOutput(
         inputs,
         'PrivateKernelTailArtifact',
-        convertPrivateKernelTailInputsToWitnessMap,
-        convertPrivateKernelTailOutputsFromWitnessMap,
+        convertPrivateKernelTailInputsToWitnessMapWithAbi,
+        convertPrivateKernelTailOutputsFromWitnessMapWithAbi,
       );
     }
     return await this.generateCircuitOutput(
       inputs,
       'PrivateKernelTailToPublicArtifact',
-      convertPrivateKernelTailToPublicInputsToWitnessMap,
-      convertPrivateKernelTailForPublicOutputsFromWitnessMap,
+      convertPrivateKernelTailToPublicInputsToWitnessMapWithAbi,
+      convertPrivateKernelTailForPublicOutputsFromWitnessMapWithAbi,
     );
   }
 
@@ -217,8 +153,12 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
     privateInputs: PrivateKernelTailCircuitPrivateInputs,
   ): Promise<PrivateKernelSimulateOutput<PrivateKernelTailCircuitPublicInputs>> {
     const isForPublic = privateInputs.isForPublic();
+    const artifactName = isForPublic ? 'PrivateKernelTailToPublicArtifact' : 'PrivateKernelTailArtifact';
+    const artifact = await this.artifactProvider.getSimulatedClientCircuitArtifactByName(artifactName);
     const [duration, result] = await elapsed(() =>
-      isForPublic ? executeTailForPublic(privateInputs) : executeTail(privateInputs),
+      isForPublic
+        ? executeTailForPublicWithArtifact(privateInputs, artifact)
+        : executeTailWithArtifact(privateInputs, artifact),
     );
     this.log.debug(`Simulated private kernel ordering`, {
       eventName: 'circuit-simulation',
@@ -233,50 +173,32 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
     );
   }
 
-  public async computeGateCountForCircuit(bytecode: Buffer, circuitName: string): Promise<number> {
-    const logFunction = (message: string) => {
-      this.log.debug(`$bb gates ${circuitName} - ${message}`);
-    };
-
-    const result = await computeGateCountForCircuit(
-      this.bbBinaryPath,
-      this.bbWorkingDirectory,
-      circuitName,
-      bytecode,
-      'mega_honk',
-      logFunction,
-    );
-    if (result.status === BB_RESULT.FAILURE) {
-      throw new Error(result.reason);
-    }
-
-    return result.circuitSize as number;
-  }
-
-  private async generateCircuitOutput<
+  public async generateCircuitOutput<
     I extends { toBuffer: () => Buffer },
     O extends PrivateKernelCircuitPublicInputs | PrivateKernelTailCircuitPublicInputs,
   >(
     inputs: I,
     circuitType: ClientProtocolArtifact,
-    convertInputs: (inputs: I) => WitnessMap,
-    convertOutputs: (outputs: WitnessMap) => O,
+    convertInputs: (inputs: I, abi: Abi) => WitnessMap,
+    convertOutputs: (outputs: WitnessMap, abi: Abi) => O,
   ): Promise<PrivateKernelSimulateOutput<O>> {
     this.log.debug(`Generating witness for ${circuitType}`);
-    const compiledCircuit: NoirCompiledCircuit = ClientCircuitArtifacts[circuitType];
+    const compiledCircuit: NoirCompiledCircuit = await this.artifactProvider.getClientCircuitArtifactByName(
+      circuitType,
+    );
 
-    const witnessMap = convertInputs(inputs);
+    const witnessMap = await convertInputs(inputs, compiledCircuit.abi);
     const timer = new Timer();
     const outputWitness = await this.simulator.simulateCircuit(witnessMap, compiledCircuit);
-    const output = convertOutputs(outputWitness);
+    const output = await convertOutputs(outputWitness, compiledCircuit.abi);
 
     this.log.debug(`Generated witness for ${circuitType}`, {
       eventName: 'circuit-witness-generation',
-      circuitName: mapProtocolArtifactNameToCircuitName(circuitType),
+      circuitName: circuitType,
       duration: timer.ms(),
       inputSize: inputs.toBuffer().length,
       outputSize: output.toBuffer().length,
-    } satisfies CircuitWitnessGenerationStats);
+    });
 
     const verificationKey = ClientCircuitVks[circuitType].keyAsFields;
     const bytecode = Buffer.from(compiledCircuit.bytecode, 'base64');
@@ -290,20 +212,7 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
     return kernelOutput;
   }
 
-  private runInDirectory<T>(fn: (dir: string) => Promise<T>) {
-    const log = this.log;
-    return runInDirectory(
-      this.bbWorkingDirectory,
-      (dir: string) =>
-        fn(dir).catch(err => {
-          log.error(`Error running operation at ${dir}: ${err}`);
-          throw err;
-        }),
-      this.skipCleanup,
-    );
-  }
-
-  private makeEmptyKernelSimulateOutput<
+  public makeEmptyKernelSimulateOutput<
     PublicInputsType extends PrivateKernelTailCircuitPublicInputs | PrivateKernelCircuitPublicInputs,
   >(publicInputs: PublicInputsType, circuitType: ClientProtocolArtifact) {
     const kernelProofOutput: PrivateKernelSimulateOutput<PublicInputsType> = {
@@ -313,5 +222,13 @@ export class BBNativePrivateKernelProver implements PrivateKernelProver {
       bytecode: Buffer.from([]),
     };
     return kernelProofOutput;
+  }
+
+  public async createClientIvcProof(acirs: Buffer[], witnessStack: WitnessMap[]): Promise<ClientIvcProof> {
+    throw new Error('Not implemented');
+  }
+
+  public computeGateCountForCircuit(_bytecode: Buffer, _circuitName: string): Promise<number> {
+    throw new Error('Not implemented');
   }
 }
