+++ yarn-project/bb-prover/src/prover/index.ts
@@ -1,3 +1,3 @@
 export * from './bb_prover.js';
-export * from './bb_private_kernel_prover.js';
+export * from './bb_native_private_kernel_prover.js';
 export * from './client_ivc_proof_utils.js';
