+++ yarn-project/bb-prover/src/wasm/bb_wasm_private_kernel_prover.ts
@@ -1,78 +1,25 @@
 import { AztecClientBackend } from '@aztec/bb.js';
-import { type PrivateKernelProver, type PrivateKernelSimulateOutput } from '@aztec/circuit-types';
-import {
-  ClientIvcProof,
-  type PrivateKernelCircuitPublicInputs,
-  type PrivateKernelInitCircuitPrivateInputs,
-  type PrivateKernelInnerCircuitPrivateInputs,
-  type PrivateKernelResetCircuitPrivateInputs,
-  type PrivateKernelTailCircuitPrivateInputs,
-  type PrivateKernelTailCircuitPublicInputs,
-} from '@aztec/circuits.js';
+import { ClientIvcProof } from '@aztec/circuits.js';
 import { createLogger } from '@aztec/foundation/log';
 import { Timer } from '@aztec/foundation/timer';
-import { type ClientProtocolArtifact } from '@aztec/noir-protocol-circuits-types/types';
-import { ClientCircuitVks } from '@aztec/noir-protocol-circuits-types/vks';
-import { WASMSimulator } from '@aztec/simulator/client';
+import { ArtifactProvider } from '@aztec/noir-protocol-circuits-types/types';
 
 import { type WitnessMap } from '@noir-lang/noir_js';
 import { serializeWitness } from '@noir-lang/noirc_abi';
 import { ungzip } from 'pako';
 
-export abstract class BBWasmPrivateKernelProver implements PrivateKernelProver {
-  protected simulator = new WASMSimulator();
+import { BBPrivateKernelProver } from '../prover/bb_private_kernel_prover.js';
 
-  constructor(protected threads: number = 1, protected log = createLogger('bb-prover:wasm')) {}
-
-  generateInitOutput(
-    _privateKernelInputsInit: PrivateKernelInitCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
+export abstract class BBWASMPrivateKernelProver extends BBPrivateKernelProver {
+  constructor(
+    protected override artifactProvider: ArtifactProvider,
+    private threads: number = 1,
+    protected override log = createLogger('bb-prover:wasm'),
+  ) {
+    super(artifactProvider, log);
   }
 
-  simulateInit(
-    _privateKernelInputsInit: PrivateKernelInitCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  generateInnerOutput(
-    _privateKernelInputsInner: PrivateKernelInnerCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  simulateInner(
-    _privateKernelInputsInner: PrivateKernelInnerCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  generateResetOutput(
-    _privateKernelInputsReset: PrivateKernelResetCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  simulateReset(
-    _privateKernelInputsReset: PrivateKernelResetCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  generateTailOutput(
-    _privateKernelInputsTail: PrivateKernelTailCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelTailCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  simulateTail(
-    _privateKernelInputsTail: PrivateKernelTailCircuitPrivateInputs,
-  ): Promise<PrivateKernelSimulateOutput<PrivateKernelTailCircuitPublicInputs>> {
-    throw new Error('Method not implemented.');
-  }
-
-  async createClientIvcProof(acirs: Buffer[], witnessStack: WitnessMap[]): Promise<ClientIvcProof> {
+  public override async createClientIvcProof(acirs: Buffer[], witnessStack: WitnessMap[]): Promise<ClientIvcProof> {
     const timer = new Timer();
     this.log.info(`Generating ClientIVC proof...`);
     const backend = new AztecClientBackend(
@@ -89,21 +36,5 @@ export abstract class BBWasmPrivateKernelProver implements PrivateKernelProver {
       vkSize: vk.length,
     });
     return new ClientIvcProof(Buffer.from(proof), Buffer.from(vk));
-  }
-
-  protected makeEmptyKernelSimulateOutput<
-    PublicInputsType extends PrivateKernelTailCircuitPublicInputs | PrivateKernelCircuitPublicInputs,
-  >(publicInputs: PublicInputsType, circuitType: ClientProtocolArtifact) {
-    const kernelProofOutput: PrivateKernelSimulateOutput<PublicInputsType> = {
-      publicInputs,
-      verificationKey: ClientCircuitVks[circuitType].keyAsFields,
-      outputWitness: new Map(),
-      bytecode: Buffer.from([]),
-    };
-    return kernelProofOutput;
-  }
-
-  computeGateCountForCircuit(_bytecode: Buffer, _circuitName: string): Promise<number> {
-    return Promise.resolve(0);
   }
 }
