+++ yarn-project/noir-protocol-circuits-types/src/artifacts/client/lazy.ts
@@ -1,14 +1,14 @@
 import { type NoirCompiledCircuit } from '@aztec/types/noir';
 
 import { ClientCircuitArtifactNames, getClientCircuitArtifact } from '../../client_artifacts_helper.js';
-import { type ClientProtocolArtifact } from '../types.js';
+import { ArtifactProvider, type ClientProtocolArtifact } from '../types.js';
 
-export function getClientCircuitArtifactByName(artifact: ClientProtocolArtifact): Promise<NoirCompiledCircuit> {
-  return getClientCircuitArtifact(ClientCircuitArtifactNames[artifact], false);
-}
+export class LazyArtifactProvider implements ArtifactProvider {
+  getClientCircuitArtifactByName(artifact: ClientProtocolArtifact): Promise<NoirCompiledCircuit> {
+    return getClientCircuitArtifact(ClientCircuitArtifactNames[artifact], false);
+  }
 
-export function getSimulatedClientCircuitArtifactByName(
-  artifact: ClientProtocolArtifact,
-): Promise<NoirCompiledCircuit> {
-  return getClientCircuitArtifact(ClientCircuitArtifactNames[artifact], true);
+  getSimulatedClientCircuitArtifactByName(artifact: ClientProtocolArtifact): Promise<NoirCompiledCircuit> {
+    return getClientCircuitArtifact(ClientCircuitArtifactNames[artifact], true);
+  }
 }
