+++ yarn-project/noir-protocol-circuits-types/src/entrypoint/client/bundle.ts
@@ -1,27 +1,5 @@
 export {
-  convertPrivateKernelInitInputsToWitnessMap,
-  convertPrivateKernelInitOutputsFromWitnessMap,
-  convertPrivateKernelInnerInputsToWitnessMap,
-  convertPrivateKernelInnerOutputsFromWitnessMap,
-  convertPrivateKernelResetInputsToWitnessMap,
-  convertPrivateKernelResetOutputsFromWitnessMap,
-  convertPrivateKernelTailForPublicOutputsFromWitnessMap,
-  convertPrivateKernelTailInputsToWitnessMap,
-  convertPrivateKernelTailOutputsFromWitnessMap,
-  convertPrivateKernelTailToPublicInputsToWitnessMap,
-  executeInit,
-  executeInner,
-  executeReset,
-  executeTail,
-  executeTailForPublic,
-} from '../../execution/client/bundle.js';
-
-export { type ClientProtocolArtifact } from '../../artifacts/types.js';
-export { ClientCircuitArtifacts } from '../../artifacts/client/bundle.js';
-
-export { getPrivateKernelResetArtifactName } from '../../utils/private_kernel_reset.js';
-export {
-  maxPrivateKernelResetDimensions,
-  privateKernelResetDimensionsConfig,
-} from '../../private_kernel_reset_types.js';
-export { foreignCallHandler } from '../../utils/client/foreign_call_handler.js';
+  BundleArtifactProvider,
+  ClientCircuitArtifacts,
+  SimulatedClientCircuitArtifacts,
+} from '../../artifacts/client/bundle.js';
