+++ yarn-project/noir-protocol-circuits-types/src/entrypoint/client/common.ts
@@ -1,0 +1,26 @@
+export {
+  convertPrivateKernelInitInputsToWitnessMapWithAbi,
+  convertPrivateKernelInitOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelInnerInputsToWitnessMapWithAbi,
+  convertPrivateKernelInnerOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelResetInputsToWitnessMapWithAbi,
+  convertPrivateKernelResetOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelTailForPublicOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelTailInputsToWitnessMapWithAbi,
+  convertPrivateKernelTailOutputsFromWitnessMapWithAbi,
+  convertPrivateKernelTailToPublicInputsToWitnessMapWithAbi,
+  executeInitWithArtifact,
+  executeInnerWithArtifact,
+  executeResetWithArtifact,
+  executeTailWithArtifact,
+  executeTailForPublicWithArtifact,
+} from '../../execution/client.js';
+
+export { getPrivateKernelResetArtifactName } from '../../utils/private_kernel_reset.js';
+export {
+  maxPrivateKernelResetDimensions,
+  privateKernelResetDimensionsConfig,
+} from '../../private_kernel_reset_types.js';
+export { foreignCallHandler } from '../../utils/client/foreign_call_handler.js';
+
+export { type ClientProtocolArtifact } from '../../artifacts/types.js';
